+++ workspaces/liferay-ticket-workspace/client-extensions/liferay-ticket-custom-element/package.json
@@ -2,13 +2,16 @@
 	"dependencies": {
 		"@clayui/alert": "^3.93.0",
 		"@clayui/icon": "^3.106.1",
+		"@clayui/layout": "^3.106.1",
 		"@clayui/loading-indicator": "^3.106.1",
 		"@clayui/nav": "^3.108.0",
 		"@clayui/pagination": "^3.94.0",
 		"@clayui/pagination-bar": "^3.94.0",
 		"@clayui/panel": "^3.106.1",
+		"@clayui/sticker": "^3.106.1",
 		"@dnd-kit/core": "^6.0.8",
 		"@yaireo/relative-time": "^1.0.3",
+		"classnames": "^2.3.2",
 		"react": "^18.2.0",
 		"react-data-grid": "^7.0.0-beta.30",
 		"react-dom": "^18.2.0",
